Update README.md
I have added a logo, links to the GitHub repository, LinkedIn and the actual link to the project to make everything easily accessible. Also, I have added detailed descriptions of the process to contribute and the tech-stacks used. Also, support section has been  created which makes it easier for the contributors to communicate with us.